requirements.txt
selenium==4.9.1
webdriver-manager==3.8.6
pytest==7.4.0
allure-pytest==2.13.2
pylint==2.17.4
flake8==6.0.0
requests==2.31.0